User input added, logic corrected
UPDATE: corrected iteration logic,  added user input which allows the user to view as many records as they like, in as much detail as they like.
This is still basic NLP and only uses word Count arrays.  The plan is to use real NLP techniques to assess more accurately if the synopsis is a match, or in other applications to find strong matches for similar texts, or see how strongly a script matches key words from the most successful movies.
The TDF-IDF logic is there, but needs to be incorporated in a meaningful way.